feat(gradient_bg): Use software rendering for GradientBg
GskGLShaderNode is a broken thing, so it isn't supported with new renderers.

The code for the software gradient rendering is adapted from the code written by Mikhail Filimonov for TelegramSwift.

The gradient is currently rendered at very low resolution (32x32) and upscaled on GPU, probably it's better to use higher resolution, but for me looks good enough.